Adapted to compile in Cygwin
git-svn-id: file:///home/roberto/Git/svn@10742 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: ac953538239ccfb006f091330e5a1805781f77f5

+++ trunk/xmipp/libraries/bindings/python/xmippmodule.cpp
@@ -2820,7 +2820,7 @@ void setMDObjectValue(MDObject *obj, PyObject *pyValue)
         if (PyInt_Check(pyValue))
             obj->setValue((int)PyInt_AS_LONG(pyValue));
         else if (PyLong_Check(pyValue))
-        	obj->setValue(PyLong_AsUnsignedLong(pyValue));
+        	obj->setValue((size_t)PyLong_AsUnsignedLong(pyValue));
         else if (PyString_Check(pyValue))
         	obj->setValue(std::string(PyString_AsString(pyValue)));
         else if (FileName_Check(pyValue))
